Move functions to FnGenerator
