+++ codegen/smithy-go-codegen/src/main/java/software/amazon/smithy/go/codegen/endpoints/FnGenerator.java
@@ -21,11 +21,15 @@ import static software.amazon.smithy.go.codegen.GoWriter.joinWritables;
 import java.util.ArrayList;
 import java.util.List;
 import software.amazon.smithy.codegen.core.Symbol;
+import software.amazon.smithy.go.codegen.GoUniverseTypes;
 import software.amazon.smithy.go.codegen.SmithyGoDependency;
 import software.amazon.smithy.go.codegen.SymbolUtils;
 import software.amazon.smithy.go.codegen.Writable;
+import software.amazon.smithy.rulesengine.language.evaluation.type.OptionalType;
 import software.amazon.smithy.rulesengine.language.syntax.expressions.Expression;
 import software.amazon.smithy.rulesengine.language.syntax.expressions.functions.FunctionDefinition;
+import software.amazon.smithy.rulesengine.language.syntax.expressions.functions.IsSet;
+import software.amazon.smithy.rulesengine.language.syntax.rule.Condition;
 import software.amazon.smithy.utils.MapUtils;
 
 public class FnGenerator {
@@ -95,6 +99,71 @@ public class FnGenerator {
 
             default -> false;
         };
+    }
+
+    /**
+     * Determine the Go type Symbol for a condition function's return value.
+     * This is used for conditionContext struct fields.
+     */
+    Symbol returnTypeForCondition(Condition condition) {
+        var fn = condition.getFunction();
+
+        // isSet stores the pointer value directly — type comes from the inner expression
+        if (fn instanceof IsSet) {
+            var inner = ((IsSet) fn).getArguments().get(0);
+            return returnTypeForExpression(inner);
+        }
+
+        // Derive Go type from the function's return type
+        var fnId = fn.getFunctionDefinition().getId();
+
+        Symbol goFn = fnProvider.fnFor(fnId);
+        if (goFn == null) {
+            goFn = new DefaultFnProvider().fnFor(fnId);
+        }
+
+        if (goFn != null) {
+            return returnTypeForFn(fnId, fn.type());
+        }
+
+        // Fallback based on Smithy type
+        var type = fn.type();
+        if (type instanceof OptionalType opt) {
+            return SymbolUtils.pointerTo(ExpressionGenerator.goTypeForType(opt.inner()));
+        }
+        return ExpressionGenerator.goTypeForType(type);
+    }
+
+    /**
+     * Map known function IDs to their Go return type Symbols.
+     */
+    static Symbol returnTypeForFn(String fnId,
+            software.amazon.smithy.rulesengine.language.evaluation.type.Type smithyType) {
+        return switch (fnId) {
+            case "parseURL" -> SymbolUtils.createPointableSymbolBuilder("URL",
+                    SmithyGoDependency.SMITHY_ENDPOINT_RULESFN).build();
+            case "substring" -> SymbolUtils.pointerTo(GoUniverseTypes.String);
+            case "aws.parseArn" -> SymbolUtils.createPointableSymbolBuilder("ARN",
+                    "github.com/aws/aws-sdk-go-v2/internal/endpoints/awsrulesfn").build();
+            case "aws.partition" -> SymbolUtils.createPointableSymbolBuilder("PartitionConfig",
+                    "github.com/aws/aws-sdk-go-v2/internal/endpoints/awsrulesfn").build();
+            case "isValidHostLabel", "aws.isVirtualHostableS3Bucket" -> GoUniverseTypes.Bool;
+            case "uriEncode" -> GoUniverseTypes.String;
+            case "split" -> SymbolUtils.sliceOf(GoUniverseTypes.String);
+            case "ite", "coalesce" -> {
+                var type = smithyType instanceof OptionalType opt ? opt.inner() : smithyType;
+                yield ExpressionGenerator.goTypeForType(type);
+            }
+            default -> GoUniverseTypes.Any;
+        };
+    }
+
+    private static Symbol returnTypeForExpression(Expression expr) {
+        var type = expr.type();
+        if (type instanceof OptionalType opt) {
+            return SymbolUtils.pointerTo(ExpressionGenerator.goTypeForType(opt.inner()));
+        }
+        return SymbolUtils.pointerTo(ExpressionGenerator.goTypeForType(type));
     }
 
 }
